Uploading Generative Mesh Trial
I had tried an example with OF before we started in class, but didn’t
understand enough about headers to get it to work. Since it featured
the easy cam in a really perfect way, I decided to follow the tutorial
again, typing it all out and explaining why everything works the way it
does. I then loaded in my own custom image for the work.